Fix libultra warnings

+++ src/libultra/os/createthread.c
@@ -28,7 +28,7 @@ void osCreateThread(OSThread* t, OSId id, void (*entry)(void*), void* arg, void*
     t->context.pc = (u32) entry;
     t->context.a0 = (s64) (s32) arg; // Double cast gets rid of compiler warning
     t->context.sp = (s64) (s32) sp - 16;
-    t->context.ra = (u64) __osCleanupThread;
+    t->context.ra = __osCleanupThread;
     mask = OS_IM_ALL;
     t->context.sr = (mask & (SR_IMASK | SR_IE)) | SR_EXL;
     t->context.rcp = (mask & RCP_IMASK) >> RCP_IMASKSHIFT;
